Increase default timeout (#84)
* Increase default timeout 

Increase default timeout to eliminate random timeouts. The default timeout is too low; the change was made from 1,000ms to 10,000ms this increase is found to be more in line with other libraries.

* Increase default timeout

Increase default timeout to eliminate random timeouts. The default timeout is too low; the change was made from 1,000ms to 10,000ms; this increase is found to be more in line with other libraries.