URL with a path and credentials is not handled correctly (#103)
This is a fix for a regression introduced by 1766e700146ee1d8da5df86b49c00689545849d3.

If a URL with a path is passed to `JsonRpcProvider`, the path is dropped when the provider makes a query because `const url = new URL(route, finalRpcUrl);` does drop the path in `finalRpcUrl`.

The proposed fix to create a URL first, and appends a path via string concat.  This patch also adds a testcase for that logic.